% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{apply_mapping}
\alias{apply_mapping}
\title{Apply dataset mapping to dataset}
\usage{
apply_mapping(dataset, mapping, validate = TRUE)
}
\arguments{
\item{dataset}{dataset to apply mapping to}

\item{mapping}{mapping to apply to dataset}

\item{validate}{whether to validate the dataset, TRUE by default}
}
\value{
\code{dataset} with renamed columns based on mapping and with attribute
"table_name" set to the expected table type and attribute "valid" indicating
if validated or not.
}
\description{
Utility function to use alongside the mapping object creator functions such as
\code{mapping_entry_points}, \code{mapping_animal_mobility}, etc.. To mapping column names
to the required values and validate the dataset for use in riskintro analysis
workflows.
}
\examples{

entry_points_fp <-
  list.files(
    system.file(
      package = "riskintrodata",
      "samples",
      "tunisia",
      "entry_points"
    ),
    full.names = TRUE
  )
entry_points <- readr::read_csv(entry_points_fp)


entry_points_mapped <- apply_mapping(
  dataset = entry_points,
  mapping = mapping_entry_points(
    point_name = "NAME",
    lng = "LONGITUDE_X",
    lat = "LATITUDE_Y",
    mode = "MODE",
    type = "TYPE",
    sources = "SOURCES"
  ),
  validate = TRUE
)

if(FALSE){
  entry_points_wrong <- entry_points
  entry_points_wrong$TYPE <- rep_len(c("PIF Maritime", "Passage contrebande", "'ELLO"),
                                     length.out = nrow(entry_points))
  entry_points_wrong$LONGITUDE_X <- rep_len(c(10.00, 15.00, NA_real_),
                                            length.out = nrow(entry_points))
  entry_points_wrong[23:30, "NAME"] <- NA_character_

  entry_points_wrong$NAME <- NULL


  entry_points_mapped <- apply_mapping(
    dataset = entry_points_wrong,
    mapping = mapping_entry_points(
      point_name = "NAME",
      lng = "LONGITUDE_X",
      lat = "LATITUDE_Y",
      mode = "MODE",
      type = "TYPE",
      sources = "SOURCES"
    ),
    validate = TRUE
  )
}


}
\seealso{
\code{\link[=mapping_entry_points]{mapping_entry_points()}}, \code{\link[=mapping_epi_units]{mapping_epi_units()}},
\code{\link[=mapping_emission_risk_factors]{mapping_emission_risk_factors()}}, \code{\link[=mapping_animal_mobility]{mapping_animal_mobility()}}
}
