% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-utils.R
\name{conn_ri}
\alias{conn_ri}
\title{Create a DuckDB Connection}
\usage{
conn_ri(location)
}
\arguments{
\item{location}{The file path to the DuckDB database. This can be a file path
to an existing DuckDB file, or a new path to create a database.}
}
\value{
A \code{DBI} connection object representing the connection to the DuckDB
database.
}
\description{
The function establishes a connection to a DuckDB database at a
specified location.
}
\examples{
# Create a connection to a DuckDB database
tf <- tempfile(fileext = ".duckdb")
con <- conn_ri(tf)
# Use the connection as needed
DBI::dbDisconnect(con, shutdown = TRUE) # Remember to disconnect when done
}
