% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{country_ref}
\alias{country_ref}
\title{Country references dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 505 rows and 3 columns.
}
\usage{
country_ref
}
\description{
Country references dataset
}
\keyword{data}
