% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso3_to_name.R
\name{country_reference}
\alias{country_reference}
\title{Table of ISO3 country codes and names in French and English}
\usage{
country_reference()
}
\value{
A data frame with columns for ISO3 codes and country names:
\itemize{
\item \code{iso3}: \emph{character} - The ISO3 country code.
\item \code{name_en}: \emph{character} - The country name in English.
\item \code{name_fr}: \emph{character} - The country name in French.
}
}
\description{
This function returns a data frame containing ISO3 country codes and their
corresponding names in languages French and English.
}
\examples{
country_reference()
}
