% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emission_risk_row.R
\name{create_emission_risk_row}
\alias{create_emission_risk_row}
\title{Create a new emission risk row}
\usage{
create_emission_risk_row(
  country_id,
  disease,
  animal_category,
  species,
  emissionRiskFactors,
  new_joker = FALSE
)
}
\arguments{
\item{country_id}{Country ID (iso3). If \code{new_joker} is TRUE, this will be
the name of a new joker ID, otherwise it should be an ISO3 code.
The ISO3 code are available in the \code{country_ref} dataset.}

\item{disease, animal_category, species}{values to be inserted}

\item{emissionRiskFactors}{table des facteurs dՎmission (reactive ou tibble)}

\item{new_joker}{logical, should be TRUE to create a new joker ID}
}
\value{
a one row tibble with the specified parameters. The dataset will
contain the following columns:
\itemize{
\item \code{iso3}: \emph{character} - The ISO3 country code or a new joker ID.
\item \code{country}: \emph{character} - The full name of the country.
\item \code{disease}: \emph{character} - The name of the disease.
\item \code{animal_category}: \emph{character} - The category of animal ("Domestic" or "Wild").
\item \code{species}: \emph{character} - The species name affected by disease.
}
}
\description{
Create a new emission risk row for a given country and disease.
This function generates a new row with the specified parameters and
returns it as a tibble.
}
