% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/road_access.R
\name{download_road_access_raster}
\alias{download_road_access_raster}
\title{Download Road Access Raster}
\source{
Weiss, D., Nelson, A., Gibson, H. et al. A global map of travel time to cities
to assess inequalities in accessibility in 2015.
Nature 553, 333–336 (2018). \url{https://doi.org/10.1038/nature25181}
}
\usage{
download_road_access_raster(force = FALSE)
}
\arguments{
\item{force}{if TRUE, the function will re-download the file even if it already exists
in the cache directory.}
}
\value{
The path to the downloaded file.
}
\description{
This function downloads a road access raster file from the 'GitHub' repository
located at \url{https://github.com/ardata-fr/road-access-raster} and saves it to
a specified destination file.

For more information on this dataset see \href{https://www.nature.com/articles/nature25181}{Nature - A global map of travel time to cities to assess inequalities in accessibility in 2015}.
}
\examples{
\dontshow{
riskintrodata_dummy_setup()
}
if (curl::has_internet()) {
  road_access_raster_file <- download_road_access_raster()
}
}
