% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{emission_risk_weights}
\alias{emission_risk_weights}
\title{Default Emission Risk Weights}
\format{
An object of class \code{list} of length 9.
}
\usage{
emission_risk_weights
}
\description{
The emission risk weights by default used to calculate emission risk scores and
emission risk from emission risk factors. Each of the measures of control and surveillence,
(see \link{wahis_emission_risk_factors}) has an associated weighting
that has been determined.
}
\keyword{data}
