% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_dataset.R
\name{extract_dataset}
\alias{extract_dataset}
\title{Extract dataset from validation status}
\usage{
extract_dataset(status)
}
\arguments{
\item{status}{A validation status object returned by \code{\link[=validate_dataset]{validate_dataset()}}.}
}
\value{
The dataset from the validation status object, with attributes
\code{table_name} and \code{valid} set.
}
\description{
Extracts the dataset from the validation status object returned by
\code{\link[=validate_dataset]{validate_dataset()}}.

If the dataset is not valid, an error is raised and the details of the
validation errors are printed.
}
\examples{
## ---- read-epi-units-tunisia ----
tun_epi_files <-
  system.file(
    package = "riskintrodata",
    "samples",
    "tunisia",
    "epi_units", "tunisia_adm2_raw.gpkg"
  )

tun_epi_unit <- read_geo_file(tun_epi_files)

validate_dataset(
  x = tun_epi_unit,
  table_name = "epi_units",
  eu_name = "NAME_2",
  user_id = "GID_2"
) |> extract_dataset()

x <- validate_dataset(
  x = tun_epi_unit,
  table_name = "epi_units",
  eu_name = "NAME_ 2",
  user_id = "GID_ 2"
)
try(extract_dataset(x))
}
