% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emission_risk_row.R
\name{get_emission_risk_edit_row}
\alias{get_emission_risk_edit_row}
\title{Get an Emission Risk Row from a Country ID}
\usage{
get_emission_risk_edit_row(
  country_id,
  emissionRiskFactors,
  disease,
  animal_category,
  species,
  new_joker = FALSE
)
}
\arguments{
\item{country_id}{an ISO3 country code (or the name of a joker)}

\item{emissionRiskFactors}{table of emission factors}

\item{disease, animal_category, species}{values to be inserted}

\item{new_joker}{if TRUE, create a new joker ID}
}
\value{
a single tibble row with the specified parameters
}
\description{
Get an emission risk row for a given country ID.
This function retrieves the row from the emissionRiskFactors dataset
and returns it as a tibble. If the row does not exist, it creates a new one.
}
