% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wahis_emission_risk_factors.R
\name{get_erf_weights}
\alias{get_erf_weights}
\title{Get Default Emission Risk Weights}
\usage{
get_erf_weights()
}
\value{
named list of emission risk weights \link{emission_risk_weights}
}
\description{
Get Default Emission Risk Weights
}
\seealso{
Other functions for "Emission Risk Factors" management: 
\code{\link{erf_row}()},
\code{\link{get_wahis_erf}()}
}
\concept{functions for "Emission Risk Factors" management}
