% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_internet}
\alias{has_internet}
\title{Check Internet Connection}
\usage{
has_internet()
}
\description{
The function tests for internet connectivity by performing a dns lookup. If a
proxy server is detected, it will also check for connectivity by connecting
via the proxy. This is a copy of \code{\link[curl:nslookup]{curl::has_internet()}}
}
\examples{
has_internet()
}
\keyword{internal}
