% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskintrodata-cache.R
\name{init_riskintrodata_cache}
\alias{init_riskintrodata_cache}
\title{Initialize 'riskintrodata' Cache}
\usage{
init_riskintrodata_cache(force = FALSE)
}
\arguments{
\item{force}{Logical. If \code{TRUE}, remove existing cache directory.}
}
\value{
The path to the 'riskintrodata' cache directory.
}
\description{
Initialize the 'riskintrodata' cache directory.
This function creates the cache directory if it does not exist.

If the \code{force} argument is set to \code{TRUE}, it will remove any existing
cache directory before creating a new one.
}
\examples{
riskintrodata_dummy_setup()
init_riskintrodata_cache()
\dontrun{
init_riskintrodata_cache(force = TRUE)
}
}
\seealso{
Other riskintrodata cache management: 
\code{\link{riskintrodata_cache_dir}()},
\code{\link{riskintrodata_cache_exists}()},
\code{\link{riskintrodata_dummy_setup}()},
\code{\link{rm_riskintrodata_cache}()}
}
\concept{riskintrodata cache management}
