% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-aa.R
\name{is.lat}
\alias{is.lat}
\title{Check if a vector is a valid latitude}
\usage{
is.lat(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
logical vector of the same length as x
}
\description{
Check if a vector is a valid latitude
}
\concept{functions for validation process}
