% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso3_to_name.R
\name{iso3_to_name}
\alias{iso3_to_name}
\title{Convert ISO3 country codes to country names}
\usage{
iso3_to_name(x, lang = "en")
}
\arguments{
\item{x}{A character vector of ISO3 country codes.}

\item{lang}{A character string specifying the language for the country names.
Default is "en" (English), alternatives include "fr" (French).}
}
\value{
A character vector of country names corresponding to the provided ISO3 codes.
}
\description{
This function converts ISO3 country codes to country names in a specified language.
}
\details{
The function uses a reference data frame \code{country_ref} that contains ISO3
codes and their corresponding country names in multiple languages. The
data come from the \code{countries} package.
}
\examples{
iso3_to_name(c("USA", "FRA", "DEU"))
}
