% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{mapping_entry_points}
\alias{mapping_entry_points}
\title{Entry points dataset mapping}
\usage{
mapping_entry_points(
  point_name,
  lng = NULL,
  lat = NULL,
  geometry = NULL,
  mode = NULL,
  type = NULL,
  sources = NULL
)
}
\arguments{
\item{point_name}{character column naming or describing the point}

\item{lng}{numeric column, latitude of point}

\item{lat}{numeric column, longitude of point (required if no geometry column present)}

\item{geometry}{sf_POINTS column if the dataset is of class \code{sf} (simple feature)}

\item{mode}{character or factor, optional, indicates whether points are contraband (C)
or non-contraband (NC). Values should be either \code{C}, \code{NC} or missing.}

\item{type}{character or factors, option, should be one of the following:
\"PIF Aerien\", \"PIF Maritime\", \"PIF terrestre\", \"Passage contrebande\",
\"Passage transhumance\" or missing". Indicates the type of transport being used.}

\item{sources}{character, optional, this is a list of all the ISO3 country codes
that animals enter from through this entry point.}
}
\value{
mapping object to be used with \code{\link[=apply_mapping]{apply_mapping()}}
}
\description{
A dataset mapping is used with \code{\link[=apply_mapping]{apply_mapping()}} to renames, select and validation
the columns in a dataset that correspond to the parameter names below.
}
