% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{neighbours_table}
\alias{neighbours_table}
\title{Neighbours table}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 634 rows and 4 columns.
}
\usage{
neighbours_table
}
\description{
A correspondence table of all countries and their neighbours.
}
\details{
Columns:
\itemize{
\item country_id: country iso3 code
\item neighbour_id: neighbouring country iso3 code
\item border_id: primary key, combination of country_id-neighbour_id
\item border_length: shared border estimated length
}
}
\keyword{data}
\keyword{internal}
