% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_emission_risk_factor_file}
\alias{read_emission_risk_factor_file}
\title{Read Emission Risk Factor File}
\usage{
read_emission_risk_factor_file(filepath)
}
\arguments{
\item{filepath}{A character string specifying the path to the text file.}
}
\description{
Reads a text file containing emission risk factors for animal diseases.
The file should be in a tabular format with specific columns (see details).
}
\details{
The text file is expected to be in a tabular format with the following columns:

Country information:
\itemize{
\item \code{iso3}: \emph{character} - A three-letter ISO 3166-1 alpha-3 country code.
\item \code{country}: \emph{character} - The full name of the country.
}

Disease information (should be the same for the whole dataset):
\itemize{
\item \code{disease}: \emph{character} - The name of the disease.
\item \code{animal_category}: \emph{character} - The category of animal ("Domestic" or "Wild").
\item \code{species}: \emph{character} - The species name affected by disease.
}

Measures of control and surveillence, all are either 1 or 0 (0 = measure inplace, 1 = no measure inplace, i.e. 1 = there is a risk, 0 = there is no risk).
\itemize{
\item \code{disease_notification}: \emph{integer} - Indicator of notification
\item \code{targeted_surveillance}: \emph{integer} - Risk factor score for targeted surveillance efforts.
\item \code{general_surveillance}: \emph{integer} - Risk factor score for general surveillance activities.
\item \code{screening}: \emph{integer} - Risk factor score for screening measures.
\item \code{precautions_at_the_borders}: \emph{integer} - Risk factor score for precautions taken at borders.
\item \code{slaughter}: \emph{integer} - Risk factor score related to the slaughter process.
\item \code{selective_killing_and_disposal}: \emph{integer} - Risk factor score for selective killing and disposal procedures.
\item \code{zoning}: \emph{integer} - Risk factor score for zoning strategies.
\item \code{official_vaccination}: \emph{integer} - Risk factor score for official vaccination programs.
}

Epidemiological status:
\itemize{
\item \code{last_outbreak_end_date}: \emph{Date} (YYYY-MM-DD) - The end date of the last outbreak.
}

Animal commerce (indicators of level of illegality of commerce)
\itemize{
\item \code{commerce_illegal}: \emph{integer} -  either 1 or 0
\item \code{commerce_legal}: \emph{integer} -  either 1 or 0
}

Other information:
\itemize{
\item \code{data_source}: \emph{character} - Reference or source of the data.
}
}
\examples{
tun_erf_file <-
  system.file(
    package = "riskintrodata",
    "samples",
    "tunisia",
    "emission_risk_factor",
    "emission_risk_factors.csv"
  )
x <- read_emission_risk_factor_file(tun_erf_file)
x
}
