% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_geo_file}
\alias{read_geo_file}
\title{Import Geo Data from Files}
\usage{
read_geo_file(x)
}
\arguments{
\item{x}{A character vector of file path(s) to the geographic data file(s).}
}
\value{
A \code{sf} object representing the imported geographic data.
}
\description{
This function imports geographic data from files. It handles shapefiles
specifically, as they consist of multiple files. The function checks for
the presence of shapefiles and reads them accordingly.
}
\details{
This function is designed to read geographic data files, it's a wrapper
around the \code{\link[sf:st_read]{sf::read_sf()}} function. It checks if the input files are
shapefiles and handles them accordingly. If the input files are not
shapefiles, it reads them directly using \code{\link[sf:st_read]{sf::read_sf()}}.
}
\examples{
tun_files <-
  list.files(
    system.file(
      package = "riskintrodata",
      "samples",
      "tunisia",
      "epi_units"
    ),
    full.names = TRUE
  )
read_geo_file(tun_files)

nga_files <- system.file(
  package = "riskintrodata",
  "samples",
  "nigeria",
  "epi_units",
  "NGA-ADM1.geojson"
)
read_geo_file(nga_files)
}
