% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_raster}
\alias{read_raster}
\title{Read Raster File}
\usage{
read_raster(x)
}
\arguments{
\item{x}{A character string specifying the path to the raster file.}
}
\value{
A \code{SpatRaster} object representing the raster data.
}
\description{
Reads a raster file from the specified path. The function checks if the
file exists and if it is a valid raster file (tif, tiff, png).
}
\examples{
if (curl::has_internet()) {
  road_access_raster_file <- download_road_access_raster()
  ra_raster <- read_raster(x = road_access_raster_file)
}
}
