% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskintrodata-cache.R
\name{riskintrodata_cache_dir}
\alias{riskintrodata_cache_dir}
\title{Get 'riskintrodata' Cache Directory}
\usage{
riskintrodata_cache_dir()
}
\description{
Get the path to the 'riskintrodata' cache directory.
This directory is used to store cached data from the 'riskintrodata' package.
}
\details{
This directory is managed by R function \code{\link[=R_user_dir]{R_user_dir()}} but can also
be defined in a non-user location by setting ENV variable \code{RISKINTRODATA_CACHE_DIR}
or by setting R option \code{RISKINTRODATA_CACHE_DIR}.
}
\examples{
riskintrodata_cache_dir()

options(RISKINTRODATA_CACHE_DIR = tempdir())
riskintrodata_cache_dir()
options(RISKINTRODATA_CACHE_DIR = NULL)

Sys.setenv(RISKINTRODATA_CACHE_DIR = tempdir())
riskintrodata_cache_dir()
Sys.setenv(RISKINTRODATA_CACHE_DIR = "")
}
\seealso{
Other riskintrodata cache management: 
\code{\link{init_riskintrodata_cache}()},
\code{\link{riskintrodata_cache_exists}()},
\code{\link{riskintrodata_dummy_setup}()},
\code{\link{rm_riskintrodata_cache}()}
}
\concept{riskintrodata cache management}
