% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskintrodata-cache.R
\name{riskintrodata_cache_exists}
\alias{riskintrodata_cache_exists}
\title{Check if 'riskintrodata' Cache Exists}
\usage{
riskintrodata_cache_exists()
}
\description{
Check if the 'riskintrodata' cache directory exists.
}
\examples{
\dontrun{
rm_riskintrodata_cache()
riskintrodata_cache_exists()
}

riskintrodata_dummy_setup()
init_riskintrodata_cache()
riskintrodata_cache_exists()
}
\seealso{
Other riskintrodata cache management: 
\code{\link{init_riskintrodata_cache}()},
\code{\link{riskintrodata_cache_dir}()},
\code{\link{riskintrodata_dummy_setup}()},
\code{\link{rm_riskintrodata_cache}()}
}
\concept{riskintrodata cache management}
\keyword{internal}
