% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskintrodata-cache.R
\name{riskintrodata_dummy_setup}
\alias{riskintrodata_dummy_setup}
\title{dummy 'riskintrodata' cache}
\usage{
riskintrodata_dummy_setup()
}
\description{
dummy 'riskintrodata' cache
used for examples or docker environments.
}
\examples{
\dontrun{
riskintrodata_dummy_setup()
riskintrodata_cache_dir()
rm_riskintrodata_cache()
}
}
\seealso{
Other riskintrodata cache management: 
\code{\link{init_riskintrodata_cache}()},
\code{\link{riskintrodata_cache_dir}()},
\code{\link{riskintrodata_cache_exists}()},
\code{\link{rm_riskintrodata_cache}()}
}
\concept{riskintrodata cache management}
