% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskintrodata-cache.R
\name{rm_riskintrodata_cache}
\alias{rm_riskintrodata_cache}
\title{Remove 'riskintrodata' Cache}
\usage{
rm_riskintrodata_cache()
}
\description{
Remove the 'riskintrodata' cache directory.
This function deletes the cache directory and all its contents.
}
\examples{
\dontrun{
riskintrodata_dummy_setup()
rm_riskintrodata_cache()
}
}
\seealso{
Other riskintrodata cache management: 
\code{\link{init_riskintrodata_cache}()},
\code{\link{riskintrodata_cache_dir}()},
\code{\link{riskintrodata_cache_exists}()},
\code{\link{riskintrodata_dummy_setup}()}
}
\concept{riskintrodata cache management}
\keyword{internal}
