% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emission_risk_row.R
\name{update_emission_risk_row_from_input}
\alias{update_emission_risk_row_from_input}
\title{Update Emission Risk Row from Input}
\usage{
update_emission_risk_row_from_input(row, input)
}
\arguments{
\item{row}{ligne tibble à modifier}

\item{input}{objet list ou liste nommée simulant \code{input$...} de Shiny}
}
\value{
tibble modifié avec colonnes binaires pour les facteurs
}
\description{
Updates an edit line with user input values
}
