% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_table_content.R
\name{validate_dataset_content_cli_msg}
\alias{validate_dataset_content_cli_msg}
\title{Format validate_dataset_content into CLI message}
\usage{
validate_dataset_content_cli_msg(x, cli = TRUE)
}
\arguments{
\item{x}{the output of \code{\link[=validate_dataset_content]{validate_dataset_content()}}}
}
\value{
the dataset of \code{x}, throws errors if there are any validation issues
from \code{\link[=validate_dataset_content]{validate_dataset_content()}}
}
\description{
Format validate_dataset_content into CLI message
}
\examples{
x <- get_wahis_erf(
  disease = "Anthrax",
  species = "Cattle",
  animal_category = "Domestic",
  validate = TRUE # default
)
status <- validate_dataset_content(x, table_name = "emission_risk_factors")
dataset <- validate_dataset_content_cli_msg(status)
dataset

x <- wahis_emission_risk_factors
status <- validate_dataset_content(x, table_name = "emission_risk_factors")
z <- try(validate_dataset_content_cli_msg(status), silent = TRUE) # error here if invalid data
message(cli::ansi_strip(z))


wrong_data <- mtcars
wrong_data$slaughter <- as.integer(round(runif(nrow(mtcars))))
wrong_data$last_outbreak_end_date <- "HELLO"
status <- validate_dataset_content(wrong_data, table_name = "emission_risk_factors")
dataset <- try(validate_dataset_content_cli_msg(status), silent = TRUE)
message(cli::ansi_strip(z))
}
