% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_table_content.R
\name{validate_table_content_cli_msg}
\alias{validate_table_content_cli_msg}
\title{Format validate_table_content into CLI message}
\usage{
validate_table_content_cli_msg(x)
}
\arguments{
\item{x}{the output of \code{\link[=validate_table_content]{validate_table_content()}}}
}
\value{
the dataset of \code{x}, throws errors if there are any validation issues
from \code{\link[=validate_table_content]{validate_table_content()}}
}
\description{
Format validate_table_content into CLI message
}
\examples{

x <- wahis_emission_risk_factors
status <- validate_table_content(x, table_name = "emission_risk_factors")
dataset <- validate_table_content_cli_msg(status) # error here if invalid data

# example of non-valid data
if(FALSE){
wrong_data <- mtcars
wrong_data$slaughter <- as.integer(round(runif(nrow(mtcars))))
wrong_data$last_outbreak_end_date <- "HELLO"
status <- validate_table_content(wrong_data, table_name = "emission_risk_factors")
dataset <- validate_table_content_cli_msg(status) # error here if invalid data
}

}
