% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{wahis_emission_risk_factors}
\alias{wahis_emission_risk_factors}
\title{Emission Risk Factors dataset from WAHIS}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16940 rows and 18 columns.
}
\usage{
wahis_emission_risk_factors
}
\description{
This data set is a combination of two datasets downloaded from the WAHIS
(\href{https://wahis.woah.org/#/home}{World Animal Health Information System}).
}
\details{
The dataset categories measures taken by governments to control and survey the
animal disease situation in any one country for a specific animal and disease.

This dataset serves as the starting point for risk of introduction analysis.
An analysis would begin by filtering the data for one type of animal species and
disease, then calculating the risk of introduction of this disease through movements
of this species.

Note that for some countries, diseases or animals - or cross section of the three -
there is no data.
\subsection{Data source}{
\enumerate{
\item \href{https://wahis.woah.org/#/dashboards/control-measure-dashboard}{Control measures dataset}
exported from the
corresponding WAHIS dashboard. Exported for the most recent time period for all countries,
diseases and species.
\item \href{https://wahis.woah.org/#/dashboards/country-or-disease-dashboard}{Disease situation dataset}
exported from the
corresponding WAHIS dashboard. Exported all data for all countries, then filtered to have
data for the most recent outbreak across each country, disease and species.
}

These two datasets are transformed and combined to create the WAHIS emission
risk factors dataset which is essential to the analysis methods in this package.
}

\subsection{Data specifications}{

The dataset contains 17 columns, each with specific types and data rules:

Country information:
\itemize{
\item \code{iso3}: \emph{character} - A three-letter ISO 3166-1 alpha-3 country code.
\item \code{country}: \emph{character} - The full name of the country.
}

Disease information (should be the same for the whole dataset):
\itemize{
\item \code{disease}: \emph{character} - The name of the disease.
\item \code{animal_category}: \emph{character} - The category of animal ("Domestic" or "Wild").
\item \code{species}: \emph{character} - The species name affected by disease.
}

Measures of control and surveillence, all are either 1 or 0 (0 = measure inplace, 1 = no measure inplace, i.e. 1 = there is a risk, 0 = there is no risk).
\itemize{
\item \code{disease_notification}: \emph{integer} - Indicator of notification
\item \code{targeted_surveillance}: \emph{integer} - Risk factor score for targeted surveillance efforts.
\item \code{general_surveillance}: \emph{integer} - Risk factor score for general surveillance activities.
\item \code{screening}: \emph{integer} - Risk factor score for screening measures.
\item \code{precautions_at_the_borders}: \emph{integer} - Risk factor score for precautions taken at borders.
\item \code{slaughter}: \emph{integer} - Risk factor score related to the slaughter process.
\item \code{selective_killing_and_disposal}: \emph{integer} - Risk factor score for selective killing and disposal procedures.
\item \code{zoning}: \emph{integer} - Risk factor score for zoning strategies.
\item \code{official_vaccination}: \emph{integer} - Risk factor score for official vaccination programs.
}

Epidemiological status:
\itemize{
\item \code{last_outbreak_end_date}: \emph{Date} (YYYY-MM-DD) - The end date of the last outbreak.
}

Animal commerce, indicators of wherther legal and illegal commerce are being conducted
between the epidemiological units and this country.
\itemize{
\item \code{commerce_illegal}: \emph{integer} -  either 1 or 0
\item \code{commerce_legal}: \emph{integer} -  either 1 or 0
}

Other information:
\itemize{
\item \code{data_source}: \emph{character} - Reference or source of the data.
}
}
}
\keyword{data}
