% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{world_sf}
\alias{world_sf}
\title{World Simple Features}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 227 rows and 4 columns.
}
\source{
\code{\link[geodata:gadm]{geodata::gadm()}}
}
\usage{
world_sf
}
\description{
SF dataset containing global administrative boundaries for most countries.
}
\keyword{data}
\keyword{internal}
