% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_table_content.R
\name{check_dataset_valid}
\alias{check_dataset_valid}
\title{Check if Dataset is Valid}
\usage{
check_dataset_valid(x, arg = deparse(substitute(x)), call = caller_env())
}
\arguments{
\item{x}{A validation status object returned by \code{\link[=validate_dataset]{validate_dataset()}}.}

\item{arg}{A character string representing the name of the argument.}

\item{call}{The environment from which the function was called.}
}
\value{
NULL, if invalid an error is raised.
}
\description{
Checks if the dataset is valid based on the validation status object
returned by \code{\link[=validate_dataset]{validate_dataset()}}.
}
\examples{
nc_data <- list.files(
    system.file("shape", package="sf"),
    full.names = TRUE,
    pattern = "^nc"
  ) |>
  read_geo_file()

nc_epi_unit <- validate_dataset(
  x = nc_data,
  table_name = "epi_units",
  eu_name = "NAME",
  user_id = "FIPS"
)

check_dataset_valid(nc_epi_unit)

nc_epi_unit <- validate_dataset(
  x = nc_data,
  table_name = "epi_units",
  eu_name = "BLAH",
  user_id = "FIPS"
)

try(check_dataset_valid(nc_epi_unit))
}
