% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/road_access.R
\name{download_road_access_raster}
\alias{download_road_access_raster}
\title{Download Road Access Raster}
\usage{
download_road_access_raster(destfile = NULL)
}
\arguments{
\item{destfile}{A character string with the name where the downloaded file
is saved. If NULL, a temporary file is created.}
}
\value{
The path to the downloaded file.
}
\description{
This function downloads a road access raster file from the 'GitHub' repository
located at \url{https://github.com/ardata-fr/road-access-raster} and saves it to
a specified destination file.
}
\examples{
if (curl::has_internet()) {
  road_access_raster_file <- download_road_access_raster()
}
}
