% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/road_access.R
\name{download_road_access_raster}
\alias{download_road_access_raster}
\title{Download Road Access Raster}
\usage{
download_road_access_raster(force = FALSE)
}
\arguments{
\item{force}{if TRUE, the function will re-download the file even if it already exists
in the cache directory.}
}
\value{
The path to the downloaded file.
}
\description{
This function downloads a road access raster file from the 'GitHub' repository
located at \url{https://github.com/ardata-fr/road-access-raster} and saves it to
a specified destination file.
}
\examples{
\dontshow{
riskintrodata_dummy_setup()
}
if (curl::has_internet()) {
  road_access_raster_file <- download_road_access_raster()
}
}
