% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_table_content.R
\name{extract_dataset}
\alias{extract_dataset}
\title{Extract dataset from validation status}
\usage{
extract_dataset(status)
}
\arguments{
\item{status}{A validation status object returned by \code{\link[=validate_dataset_content]{validate_dataset_content()}}.}
}
\value{
The dataset from the validation status object, with attributes
\code{table_name} and \code{valid} set.
}
\description{
Extracts the dataset from the validation status object returned by
\code{\link[=validate_dataset_content]{validate_dataset_content()}}.
}
\examples{
## ---- read-epi-units-tunisia ----
tun_epi_files <-
  system.file(
    package = "riskintrodata",
    "samples",
    "tunisia",
    "epi_units", "tunisia_adm2_raw.gpkg"
  )

tun_epi_unit <- read_geo_file(tun_epi_files)

DATA_EPI_UNITS <- validate_dataset_content(
  x = tun_epi_unit,
  table_name = "epi_units",
  eu_name = "NAME_2",
  user_id = "GID_2"
)

dat <- extract_dataset(DATA_EPI_UNITS)
# dat
# attr(dat, "table_name")
# attr(dat, "valid")
# check_dataset_valid(dat)
}
