% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wahis_emission_risk_factors.R
\name{get_wahis_erf}
\alias{get_wahis_erf}
\title{Get WAHIS Emission Risk Factors Dataset}
\usage{
get_wahis_erf(
  disease = character(),
  species = character(),
  animal_category = character()
)
}
\arguments{
\item{disease}{filter dataset for one or more disease}

\item{species}{filter dataset for one or more species}

\item{animal_category}{filter dataset for one or more animal_category}
}
\value{
the emission risk factorts dataset as documented here \link{wahis_emission_risk_factors}
}
\description{
Helper function for getting the WAHIS emission risk factors dataset. As most
analysis done require filtering for one type of each of diease, species and
animal_category, this function is a helper for that.

For dataset documentation see: \link{wahis_emission_risk_factors}.

The function is supposed to be used to get data for all countries that could
be involved in the introduction of a risk. The function can not be used to
gather data for different diseases, species or animal categories at the same time.
}
\examples{
get_wahis_erf(
  disease = "Anthrax",
  species = "Cattle",
  animal_category = "Domestic"
)
}
\seealso{
Other functions for "Emission Risk Factors" management: 
\code{\link{erf_row}()},
\code{\link{get_erf_weights}()}
}
\concept{functions for "Emission Risk Factors" management}
