% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wahis_emission_risk_factors.R
\name{get_wahis_erf}
\alias{get_wahis_erf}
\title{Get WAHIS Emission Risk Factors Dataset}
\usage{
get_wahis_erf(
  disease,
  species,
  animal_category = c("Domestic", "Wild"),
  quiet = FALSE
)
}
\arguments{
\item{disease}{filter dataset for one or more disease}

\item{species}{filter dataset for one or more species}

\item{animal_category}{filter dataset for one or more animal_category}

\item{quiet}{whether to quieten information message. Default is FALSE.}
}
\value{
A tibble of emission risk factors data. Also has the attributes:
\itemize{
\item \code{table_name = "emission_risk_factors"}
\item \code{study_settings} which is a named character vector containing the
parameter arguments.
}
}
\description{
Helper function for getting the WAHIS emission risk factors dataset. Riskintro
analysis is intended one value of each of \code{disease}, \code{species} and
\code{animal_category}, this function is a helper for that. That is to say only
one disease in one animal is analysed at a time.

For dataset documentation see: \link{wahis_emission_risk_factors}
}
\examples{
data_for_riskintro_study <- get_wahis_erf(
  disease = "Anthrax",
  species = "Cattle",
  animal_category = "Domestic"
)
data_for_riskintro_study
attr(data_for_riskintro_study, "study_settings")
}
\seealso{
Other functions for "Emission Risk Factors" management: 
\code{\link{erf_row}()},
\code{\link{get_erf_weights}()}
}
\concept{functions for "Emission Risk Factors" management}
