% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec-aa.R
\name{is.lng}
\alias{is.lng}
\title{Check if a vector is a valid longitude}
\usage{
is.lng(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
logical vector of the same length as x
}
\description{
Check if a vector is a valid longitude
}
\seealso{
Other functions for validation process: 
\code{\link{is.lat}()}
}
\concept{functions for validation process}
\keyword{internal}
