% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_table_content.R
\name{is_dataset_valid}
\alias{is_dataset_valid}
\title{Check if Dataset is Valid}
\usage{
is_dataset_valid(x)
}
\arguments{
\item{x}{A validation status object returned by \code{\link[=validate_dataset_content]{validate_dataset_content()}}.}
}
\value{
A logical value indicating whether the dataset is valid.
}
\description{
Checks if the dataset is valid based on the validation status object
returned by \code{\link[=validate_dataset_content]{validate_dataset_content()}}.
}
\examples{
library(riskintrodata)
tun_epi_files <-
  system.file(
    package = "riskintrodata",
    "samples",
    "tunisia",
    "epi_units", "tunisia_adm2_raw.gpkg"
  )

tun_epi_unit <- read_geo_file(tun_epi_files)

DATA_EPI_UNITS <- validate_dataset_content(
  x = tun_epi_unit,
  table_name = "epi_units",
  eu_name = "NAME_2",
  user_id = "GID_2"
)

is_dataset_valid(DATA_EPI_UNITS)
}
