% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{mapping_animal_mobility}
\alias{mapping_animal_mobility}
\title{Animal movement dataset mapping}
\usage{
mapping_animal_mobility(
  o_iso3 = NULL,
  o_name,
  o_lng = NULL,
  o_lat = NULL,
  d_iso3 = NULL,
  d_name,
  d_lng,
  d_lat,
  quantity = NULL
)
}
\arguments{
\item{o_iso3}{character, optional,  origin country iso3 code}

\item{o_name}{character, reauired, origin name or description}

\item{o_lng}{numeric, optional, origin point longitude}

\item{o_lat}{numeric, option, origin point latitude}

\item{d_iso3}{character, optional,  destination country iso3 code}

\item{d_name}{character, origin name or description}

\item{d_lng}{numeric, required, destination point longitude}

\item{d_lat}{numeric, required, destination point latitude}

\item{quantity}{numeric, optional, used to weight animal movement flows by quantity
of animals, if not provided no weighting is done.}
}
\value{
mapping object to be used with \code{\link[=apply_mapping]{apply_mapping()}}
}
\description{
A dataset mapping is used with \code{\link[=apply_mapping]{apply_mapping()}} to renames, select and validation
the columns in a dataset that correspond to the parameter names below.
}
