% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{mapping_emission_risk_factors}
\alias{mapping_emission_risk_factors}
\title{Emission risk factors dataset mapping}
\usage{
mapping_emission_risk_factors(
  iso3,
  country,
  disease,
  animal_category,
  species,
  disease_notification,
  targeted_surveillance,
  general_surveillance,
  screening,
  precautions_at_the_borders,
  slaughter,
  selective_killing_and_disposal,
  zoning,
  official_vaccination,
  last_outbreak_end_date,
  commerce_illegal,
  commerce_legal,
  data_source = NULL
)
}
\arguments{
\item{iso3}{character, required, country iso3 code}

\item{country}{character, required, country name}

\item{disease}{character, required, disease name}

\item{animal_category}{character, required, animal category "Wild" or "Domestic"}

\item{species}{character, required, animal species name}

\item{disease_notification}{integer, required, must be NA, 1 or 0: 1 representing this policy is not
enforced in that country (i.e. 1 is risky, 0 is not).}

\item{targeted_surveillance}{as \code{disease_notification}}

\item{general_surveillance}{as \code{disease_notification}}

\item{screening}{as \code{disease_notification}}

\item{precautions_at_the_borders}{as \code{disease_notification}}

\item{slaughter}{as \code{disease_notification}}

\item{selective_killing_and_disposal}{as \code{disease_notification}}

\item{zoning}{as \code{disease_notification}}

\item{official_vaccination}{as \code{disease_notification}}

\item{last_outbreak_end_date}{date, required, must be a date: represents the
last time there was an outbreak of this disease in this country.}

\item{commerce_illegal}{integer, required, must be NA, 1 or 0: 1 representing
this type of commerce is conducted between this country and country for which
risk of introduction is been calcalated (i.e. 1 is risky, 0 is not).}

\item{commerce_legal}{as \code{commerce_illegal}}

\item{data_source}{character, optional, used to keep track of where the data
is sourced.}
}
\value{
mapping object to be used with \code{\link[=apply_mapping]{apply_mapping()}}
}
\description{
A dataset mapping is used with \code{\link[=apply_mapping]{apply_mapping()}} to renames, select and validation
the columns in a dataset that correspond to the parameter names below.
}
\details{
For more about the data requirements of this dataset see \link{wahis_emission_risk_factors} documentation
}
