% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{mapping_entry_points}
\alias{mapping_entry_points}
\title{Entry Points Mapping}
\usage{
mapping_entry_points(
  point_name,
  lng = NULL,
  lat = NULL,
  geometry = NULL,
  mode = NULL,
  type = NULL,
  sources = NULL
)
}
\arguments{
\item{point_name}{character column naming or describing the point}

\item{lng}{numeric column, latitude of point}

\item{lat}{numeric column, longitude of point (required if no geometry column present)}

\item{geometry}{sf_POINTS column if the dataset is of class \code{sf} (simple feature)}

\item{mode}{character or factor, optional, indicates whether points are contraband (C)
or non-contraband (NC). Values should be either \code{C}, \code{NC} or missing.}

\item{type}{Optional, indicates the type of transport being used,
this is for displaying on maps and does not affect risk analysis scores. It
can be a character or a factor, should be one of the following:
\itemize{
\item "AIR" (airport),
\item "SEA" (sea port),
\item "BC" (border crossing),
\item "CC" (contraband crossing),
\item "TC" (transhumance crossing)
\item NA (missing)".
}}

\item{sources}{character, optional, this is a list of all the ISO3 country codes
that animals enter from through this entry point.}
}
\value{
mapping object to be used by \code{\link[=validate_dataset]{validate_dataset()}}
}
\description{
Get mapping description for \emph{Entry Points}.
}
\details{
A dataset mapping is used by \code{\link[=validate_dataset]{validate_dataset()}} to rename, select and validation
the columns in a dataset that correspond to the parameter names below.
}
\seealso{
Other functions for mapping tables: 
\code{\link{mapping_animal_mobility}()},
\code{\link{mapping_emission_risk_factors}()},
\code{\link{mapping_epi_units}()}
}
\concept{functions for mapping tables}
