% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{mapping_epi_units}
\alias{mapping_epi_units}
\title{Epidemiological units dataset mapping}
\usage{
mapping_epi_units(eu_name, eu_id = NULL, user_id = NULL, geometry)
}
\arguments{
\item{eu_name}{character, required, name or description of epi units}

\item{eu_id}{character, optional, epi unit identifier column}

\item{user_id}{charcter, optional, can be provided to join other datasets}

\item{geometry}{sf_POLYGON or sf_MULTIPOLYGON geospatial data type, representing
the geographical areas of each epi unit.}
}
\value{
mapping object to be used with \code{\link[=apply_mapping]{apply_mapping()}}
}
\description{
A dataset mapping is used with \code{\link[=apply_mapping]{apply_mapping()}} to renames, select and validation
the columns in a dataset that correspond to the parameter names below.
}
