% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{mapping_epi_units}
\alias{mapping_epi_units}
\title{Epidemiological Units Mapping}
\usage{
mapping_epi_units(eu_id = NULL, eu_name, geometry)
}
\arguments{
\item{eu_id}{charcter, optional, can be provided to join other datasets if needed}

\item{eu_name}{character, required, name or description of epi units}

\item{geometry}{sf_POLYGON or sf_MULTIPOLYGON geospatial data type, representing
the geographical areas of each epi unit.}
}
\value{
mapping object to be used by \code{\link[=validate_dataset_content]{validate_dataset_content()}}
}
\description{
Get mapping description for \emph{Epidemiological Units}.
}
\details{
A dataset mapping is used by \code{\link[=validate_dataset_content]{validate_dataset_content()}} to rename, select and validation
the columns in a dataset that correspond to the parameter names below.
}
\seealso{
Other functions for mapping tables: 
\code{\link{mapping_animal_mobility}()},
\code{\link{mapping_emission_risk_factors}()},
\code{\link{mapping_entry_points}()}
}
\concept{functions for mapping tables}
