% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_dataset.R
\name{validate_dataset}
\alias{validate_dataset}
\title{Validate 'riskintro' Datasets}
\usage{
validate_dataset(x, table_name, ...)
}
\arguments{
\item{x}{A data frame or an 'sf' object to be validated.}

\item{table_name}{A character string specifying the name of the dataset. It
accepts one of the following values:
\itemize{
\item "animal_mobility"
\item "epi_units"
\item "entry_points"
\item "emission_risk_factors"
}}

\item{...}{Additional arguments to be passed to the function. It is expected
to be a named list of columns to be renamed in the dataset.
For example, \verb{col1 = "new_col1", col2 = "new_col2"}.}
}
\value{
A list containing the validation status of the dataset. The list
contains the following elements:
\itemize{
\item \code{required_columns}: A list with the status of required columns.
\item \code{optional_columns}: A list with the status of optional columns.
\item \code{validate_rules}: A list with the status of validation rules.
\item \code{dataset}: The dataset after renaming and selecting the specified columns.
If validation fails, this element will be NULL, not available. You should
use \code{\link[=extract_dataset]{extract_dataset()}} to extract the dataset from the validation status object.
}
}
\description{
Validates the datasets with 'riskintro' rules.
The function checks the required and optional columns,
and validates the data using the rules defined in the specifications.
}
\details{
The function checks if the dataset contains the required and optional columns
as specified in the specifications. It also validates the data using the
rules defined in the specifications. If errors are met or validity rules
are not satisfied, the function returns a list with the status of the validation.
}
\examples{
## ---- read-epi-units-tunisia ----
tun_epi_files <-
  system.file(
    package = "riskintrodata",
    "samples",
    "tunisia",
    "epi_units", "tunisia_adm2_raw.gpkg"
  )

tun_epi_unit <- read_geo_file(tun_epi_files)

DATA_EPI_UNITS <- validate_dataset(
  x = tun_epi_unit,
  table_name = "epi_units",
  eu_name = "shapeName",
  user_id = "fid"
)

DATA_EPI_UNITS
## ---- read-animal-mobility-tunisia ----
tun_animal_mobility <-
  system.file(
    package = "riskintrodata",
    "samples",
    "tunisia",
    "animal_mobility", "ANIMAL_MOBILITY_raw.csv"
  )

x <- readr::read_csv(
  tun_animal_mobility,
)

DATA_ANIMAL_MOBILITY <- validate_dataset(
  x,
  table_name = "animal_mobility",
  o_name = "ORIGIN_NAME",
  o_lng = "ORIGIN_LONGITUDE_X",
  o_lat = "ORIGIN_LATITUDE_Y",
  d_name = "DESTINATION_NAME",
  d_lng = "DESTINATION_LONGITUDE_X",
  d_lat = "DESTINATION_LATITUDE_Y",
  quantity = "HEADCOUNT"
)

extract_dataset(DATA_ANIMAL_MOBILITY)
# read Tunisia emission risk factors dataset ----
tun_erf_file <-
  system.file(
    package = "riskintrodata",
    "samples",
    "tunisia",
    "emission_risk_factor",
    "emission_risk_factors.csv"
  )
x <- read_emission_risk_factor_file(tun_erf_file)

DATA_ERF <- validate_dataset(x, table_name = "emission_risk_factors")

DATA_ERF
# read entry points data from Tunisia ----
tun_entry_points <- system.file(
  package = "riskintrodata",
  "samples",
  "tunisia",
  "entry_points",
  "BORDER_CROSSING_POINTS.csv"
)

x <- readr::read_delim(
  tun_entry_points
)

DATA_ENTRY_POINTS <- validate_dataset(
  x = x,
  table_name = "entry_points",
  point_name = "NAME",
  lng = "LONGITUDE_X",
  lat = "LATITUDE_Y",
  mode = "MODE",
  type = "TYPE"
)

DATA_ENTRY_POINTS

DATA_ENTRY_POINTS <- validate_dataset(
  x = x,
  table_name = "entry_points",
  point_name = "NAME",
  lng = "LONGITUDE_X",
  lat = "LATITUDE_Y",
  mode = "MODE",
  type = "TYPE",
  sources = "SOURCES"
)
extract_dataset(DATA_ENTRY_POINTS)





entry_points_wrong <- x
entry_points_wrong$TYPE <- rep_len(c("PIF Maritime", "Passage contrebande", "'ELLO"),
                                   length.out = nrow(x))
entry_points_wrong$LONGITUDE_X <- rep_len(c(10.00, 15.00, NA_real_),
                                          length.out = nrow(x))
entry_points_wrong[23:30, "NAME"] <- NA_character_

entry_points_wrong$NAME <- NULL

entry_points_mapped <- validate_dataset(
  x = entry_points_wrong,
  table_name = "entry_points",
  point_name = "NAME",
  lng = "LONGITUDE_X",
  lat = "LATITUDE_Y",
  mode = "MODE",
  type = "TYPE",
  sources = "SOURCES"
)
try(extract_dataset(entry_points_mapped))
## ---- read-epi-units-tunisia ----
tun_epi_files <-
  system.file(
    package = "riskintrodata",
    "samples",
    "tunisia",
    "epi_units", "tunisia_adm2_raw.gpkg"
  )

tun_epi_unit <- read_geo_file(tun_epi_files)

DATA_EPI_UNITS <- validate_dataset(
  x = tun_epi_unit,
  table_name = "epi_units",
  eu_name = "shape Name",
  user_id = "fid"
)

try(extract_dataset(DATA_EPI_UNITS))
}
